# docs/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Red Light Green Light Platform API",
    "version": "1.0.0",
    "description": "Unified public API served through the gateway. Generated from per-service controller annotations — do not edit by hand (run `npm run generate:openapi`).",
    "contact": {
      "name": "Red Light Green Light Support"
    }
  },
  "servers": [
    {
      "url": "/",
      "description": "API Gateway"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "TestResult": {
        "properties": {
          "comment": {
            "type": "string"
          },
          "id": {
            "type": "integer"
          },
          "status": {
            "enum": ["passed", "failed", "blocked", "skipped", "untested"],
            "type": "string"
          }
        },
        "type": "object"
      },
      "TestRun": {
        "properties": {
          "description": {
            "type": "string"
          },
          "id": {
            "type": "integer"
          },
          "name": {
            "type": "string"
          },
          "projectId": {
            "type": "integer"
          },
          "status": {
            "enum": ["open", "closed"],
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "UpdateResultInput": {
        "properties": {
          "comment": {
            "maxLength": 1000,
            "type": "string"
          },
          "status": {
            "enum": ["Passed", "Failed", "Blocked", "Untested"],
            "type": "string"
          },
          "version": {
            "description": "Optimistic-concurrency token; update fails with 409 if stale",
            "type": "integer"
          }
        },
        "required": ["status"],
        "type": "object"
      },
      "UpdateTestRunInput": {
        "properties": {
          "description": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "version": {
            "description": "Optimistic-concurrency token; update fails with 409 if stale",
            "type": "integer"
          }
        },
        "type": "object"
      },
...
    }
  },
  "security": [
    {
      "BearerAuth": []
    }
  ],
  "paths": {
...
    "/api/v1/testruns/results/{resultId}": {
      "put": {
        "parameters": [
          {
            "in": "path",
            "name": "resultId",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "x-project-id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateResultInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TestResult"
                }
              }
            },
            "description": "Updated test result"
          },
          "404": {
            "description": "Test result not found"
          },
          "409": {
            "description": "Version conflict (optimistic concurrency)"
          }
        },
        "summary": "Record or update the outcome of a single test result",
        "tags": ["TestRuns"]
      }
    },
    "/api/v1/testruns/{id}": {
      "delete": {
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "x-project-id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Deleted"
          },
          "404": {
            "description": "Test run not found"
          }
        },
        "summary": "Delete a test run",
        "tags": ["TestRuns"]
      },
      "get": {
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "x-project-id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TestRun"
                }
              }
            },
            "description": "Test run"
          },
          "404": {
            "description": "Test run not found"
          }
        },
        "summary": "Get a single test run (with its results) by ID",
        "tags": ["TestRuns"]
      },
      "put": {
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "x-project-id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTestRunInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Updated test run"
          },
          "404": {
            "description": "Test run not found"
          },
          "409": {
            "description": "Version conflict (optimistic concurrency)"
          }
        },
        "summary": "Update a test run's metadata",
        "tags": ["TestRuns"]
      }
    }
  }
}